+++ src/renderer2d.rs
@@ -28,11 +28,11 @@ use {
 use crate::color::Color;
 use crate::dimen::{UVec2, Vec2};
 use crate::error::{BacktraceError, Context, ErrorMessage};
-use crate::font::FormattedTextBlock;
+use crate::font::{FormattedGlyph, FormattedTextBlock};
 use crate::font_cache::GlyphCache;
 use crate::glwrapper::*;
 use crate::image::{ImageDataType, ImageHandle, ImageSmoothingMode};
-use crate::{Polygon, RawBitmapData, Rectangle};
+use crate::{Polygon, RawBitmapData, Rect, Rectangle};
 
 struct AttributeBuffers
 {
@@ -289,6 +289,15 @@ enum RenderQueueItem
         block: Rc<FormattedTextBlock>
     },
 
+    #[allow(dead_code)]
+    FormattedTextGlyph
+    {
+        position: Vec2,
+        color: Color,
+        glyph: FormattedGlyph,
+        crop_window: Rect
+    },
+
     CircleSectionColored
     {
         vertex_positions_clockwise: [Vec2; 3],
@@ -328,10 +337,26 @@ impl RenderQueueItem
             } => {
                 for line in block.iter_lines() {
                     for glyph in line.iter_glyphs() {
-                        glyph_cache
-                            .get_renderer2d_actions(glyph, *position, *color, runner);
+                        glyph_cache.get_renderer2d_actions(
+                            glyph, *position, *color, None, runner
+                        );
                     }
                 }
+            }
+
+            RenderQueueItem::FormattedTextGlyph {
+                glyph,
+                position,
+                color,
+                crop_window
+            } => {
+                glyph_cache.get_renderer2d_actions(
+                    glyph,
+                    *position,
+                    *color,
+                    Some(crop_window),
+                    runner
+                );
             }
 
             RenderQueueItem::CircleSectionColored {
@@ -557,18 +582,32 @@ impl Renderer2D
         let mut has_text = false;
 
         for item in &self.render_queue {
-            if let RenderQueueItem::FormattedTextBlock {
-                block, position, ..
-            } = item
-            {
-                for line in block.iter_lines() {
-                    for glyph in line.iter_glyphs() {
-                        self.glyph_cache
-                            .add_to_cache(&self.context, glyph, *position);
+            match item {
+                RenderQueueItem::FormattedTextBlock {
+                    block, position, ..
+                } => {
+                    for line in block.iter_lines() {
+                        for glyph in line.iter_glyphs() {
+                            self.glyph_cache.add_to_cache(
+                                &self.context,
+                                glyph,
+                                *position
+                            );
+                        }
                     }
-                }
 
-                has_text = true;
+                    has_text = true;
+                }
+                RenderQueueItem::FormattedTextGlyph {
+                    glyph, position, ..
+                } => {
+                    self.glyph_cache
+                        .add_to_cache(&self.context, glyph, *position);
+                    has_text = true;
+                }
+                RenderQueueItem::CircleSectionColored { .. }
+                | RenderQueueItem::TriangleColored { .. }
+                | RenderQueueItem::TriangleTextured { .. } => {}
             }
         }
 
@@ -836,6 +875,34 @@ impl Renderer2D
             color,
             block: text.clone()
         })
+    }
+
+    #[inline]
+    pub(crate) fn draw_text_cropped<V: Into<Vec2>>(
+        &mut self,
+        position: V,
+        crop_window: Rect,
+        color: Color,
+        text: &Rc<FormattedTextBlock>
+    )
+    {
+        let position = position.into();
+
+        for line in text.iter_lines() {
+            for glyph in line.iter_glyphs() {
+                if let Some(glyph_outline) = glyph.pixel_bounding_box() {
+                    let glyph_outline = glyph_outline.with_offset(position);
+                    if glyph_outline.intersect(&crop_window).is_some() {
+                        self.add_to_render_queue(RenderQueueItem::FormattedTextGlyph {
+                            position,
+                            color,
+                            glyph: glyph.clone(),
+                            crop_window: crop_window.clone()
+                        })
+                    }
+                }
+            }
+        }
     }
 
     #[inline]